moved ml enum classes out of cudnn directory
